Suppress a bunch of grpc wrapper and resolver frames from tracebacks
