+++ modal/_traceback.py
@@ -12,6 +12,7 @@ import warnings
 from types import TracebackType
 from typing import Any, Iterable, Optional
 
+from modal.config import config
 from modal_proto import api_pb2
 
 from ._vendor.tblib import Traceback as TBLibTraceback
@@ -115,7 +116,7 @@ def traceback_contains_remote_call(tb: Optional[TracebackType]) -> bool:
 def print_exception(exc: Optional[type[BaseException]], value: Optional[BaseException], tb: Optional[TracebackType]):
     """Add backwards compatibility for printing exceptions with "notes" for Python<3.11."""
     traceback.print_exception(exc, value, tb)
-    if sys.version_info < (3, 11) and value is not None:
+    if sys.version_info < (3, 11) and value is not None:  # type: ignore
         notes = getattr(value, "__notes__", [])
         print(*notes, sep="\n", file=sys.stderr)
 
@@ -127,3 +128,30 @@ def print_server_warnings(server_warnings: Iterable[api_pb2.Warning]):
     """
     for warning in server_warnings:
         warnings.warn_explicit(warning.message, ServerWarning, "<modal-server>", 0)
+
+
+class suppress_tb_frames:
+    def __init__(self, n: int):
+        self.n = n
+
+    def __enter__(self):
+        pass
+
+    def __exit__(self, exc_type: type[BaseException], exc: BaseException, traceback: TracebackType) -> bool:
+        if config.get("traceback"):
+            return False
+        # modify traceback on exception object
+        final_tb = traceback
+        for _ in range(self.n):
+            final_tb = final_tb.tb_next
+
+        # for some reason, the traceback cleanup here can't be moved into a context manager :(
+        traceback_suppression_note = (
+            "Internal Modal traceback frames are suppressed for readability. "
+            "Use MODAL_TRACEBACK=1 to show a full traceback."
+        )
+        exc.with_traceback(final_tb)
+        notes = getattr(exc, "__notes__", [])
+        if traceback_suppression_note not in notes:
+            exc.add_note(traceback_suppression_note)
+        return False
